+++ src/solidlsp/language_servers/scala_language_server.py
@@ -532,7 +532,7 @@ class ScalaLanguageServer(SolidLanguageServer):
             config,
             repository_root_path,
             ProcessLaunchInfo(cmd=scala_lsp_executable_path, cwd=repository_root_path),
-            config.ls_id.value,
+            "scala",
             solidlsp_settings,
         )